pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
packages = []
package-mode = false
requires-python = ">=3.12,<4.0"
dependencies = [
    "django-extensions (>=4.1,<5.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "django (>=5.1,<5.2)",
    "celery (>=5.6.0,<6.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "python-telegram-bot (>=22.5,<23.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "coverage (>=7.13.0,<8.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
